Separate CLI logic from game, more initialization cleanup
